moves device.class calculation out of contexts and into tags
