+++ relay-general/src/protocol/contexts/device.rs
@@ -2,29 +2,6 @@ use crate::protocol::FromUserAgentInfo;
 use crate::store::user_agent::is_known;
 use crate::types::{Annotated, Object, Value};
 use crate::user_agent::{parse_device, ClientHints};
-use serde::{Deserialize, Serialize};
-
-#[derive(
-    Clone,
-    Copy,
-    Debug,
-    Deserialize,
-    Eq,
-    PartialEq,
-    Serialize,
-    Empty,
-    FromValue,
-    IntoValue,
-    ProcessValue,
-)]
-#[cfg_attr(feature = "jsonschema", derive(JsonSchema))]
-pub struct DeviceClass(pub u64);
-
-impl DeviceClass {
-    pub const LOW: Self = Self(1);
-    pub const MEDIUM: Self = Self(2);
-    pub const HIGH: Self = Self(3);
-}
 
 /// Device information.
 ///
@@ -183,10 +160,6 @@ pub struct DeviceContext {
     /// Whether location support is available on the device.
     pub supports_location_service: Annotated<bool>,
 
-    // The performance class of the device, stored as a number.
-    // This value is synthesized from the device's specs in normalize_device_context.
-    pub class: Annotated<DeviceClass>,
-
     /// Additional arbitrary fields for forwards compatibility
     #[metastructure(additional_properties, retain = "true", pii = "maybe")]
     pub other: Object<Value>,
@@ -231,7 +204,6 @@ impl FromUserAgentInfo for DeviceContext {
 
 #[cfg(test)]
 mod tests {
-    use crate::protocol::contexts::device::DeviceClass;
     use crate::protocol::{DeviceContext, FromUserAgentInfo};
     use crate::protocol::{Headers, PairList};
     use crate::types::{Annotated, Object, Value};
@@ -388,7 +360,6 @@ mod tests {
             supports_gyroscope: Annotated::new(true),
             supports_audio: Annotated::new(true),
             supports_location_service: Annotated::new(true),
-            class: Annotated::new(DeviceClass::LOW),
             other: {
                 let mut map = Object::new();
                 map.insert(
