+++ relay-general/src/protocol/mod.rs
@@ -7,6 +7,7 @@ mod clientsdk;
 mod constants;
 mod contexts;
 mod debugmeta;
+mod device_class;
 mod event;
 mod exception;
 mod fingerprint;
@@ -41,6 +42,7 @@ pub use self::clientsdk::*;
 pub use self::constants::*;
 pub use self::contexts::*;
 pub use self::debugmeta::*;
+pub use self::device_class::*;
 pub use self::event::*;
 pub use self::exception::*;
 pub use self::fingerprint::*;
